backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "College Room Booking System API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "sqlalchemy[asyncio]>=2.0.41",
    "aiomysql>=0.2.0",
    "alembic>=1.16.1",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "httpx>=0.28.1",
    "ruff>=0.11.12",
    "mypy>=1.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]